Iot connect rework (#19)
Altered the iotc_connect function to take the standard MQTT credentials of username, password, and client_id. It is now up to the application to feed in the correct parameters for these values.

Please see b/122958370 for design reasons.

Work included the update of the various examples as well as the library itself.