Update iotc_types.h
There were enumerated types being tagged as return values or as parameters, which they are neither. Updated the doxygen comments to match.  Also this was written as though it's a BSP header, which it is not, so I removed the comments asking for readers to implement these functions or types.